+++ client/src/ts/modules/hangout/settings/hangoutSettings.ts
@@ -1,12 +1,14 @@
-import { dayMilliseconds, HANGOUT_AVAILABILITY_STAGE, HANGOUT_SUGGESTIONS_STAGE, HANGOUT_VOTING_STAGE, MAX_HANGOUT_MEMBERS_LIMIT, MAX_HANGOUT_PERIOD_DAYS, MIN_HANGOUT_MEMBERS_LIMIT, MIN_HANGOUT_PERIOD_DAYS } from "../../global/clientConstants";
+import { dayMilliseconds, HANGOUT_AVAILABILITY_STAGE, HANGOUT_SUGGESTIONS_STAGE, HANGOUT_VOTING_STAGE, MAX_HANGOUT_MEMBERS_LIMIT, MAX_HANGOUT_PERIOD_DAYS, MIN_HANGOUT_MEMBERS_LIMIT } from "../../global/clientConstants";
 import LoadingModal from "../../global/LoadingModal";
 import popup from "../../global/popup";
 import SliderInput from "../../global/SliderInput";
 import { globalHangoutState } from "../globalHangoutState";
-import { directlyNavigateHangoutSections } from "../hangoutNav";
+import { directlyNavigateHangoutSections, navigateHangoutSections } from "../hangoutNav";
+import { calculateStepMinimumSliderValue, resetMembersLimitSliderValues, resetSliderValues, resetStageSliderValues, updateSettingsButtons } from "./settingsUtils";
 
 interface HangoutSettingsState {
   isLoaded: boolean,
+  settingsSectionMutationObserverActive: boolean,
 
   unsavedStageChanges: boolean,
   unsavedMembersLimitChanges: boolean,
@@ -20,8 +22,9 @@ interface HangoutSettingsState {
   },
 };
 
-const hangoutSettingsState: HangoutSettingsState = {
+export const hangoutSettingsState: HangoutSettingsState = {
   isLoaded: false,
+  settingsSectionMutationObserverActive: false,
 
   unsavedStageChanges: false,
   unsavedMembersLimitChanges: false,
@@ -30,12 +33,18 @@ const hangoutSettingsState: HangoutSettingsState = {
   sliders: null,
 };
 
+const settingsSectionElement: HTMLElement | null = document.querySelector('#settings-section');
+const updateHangoutPasswordForm: HTMLFormElement | null = document.querySelector('#hangout-settings-password-form');
+
 export function hangoutSettings(): void {
   loadEventListeners();
 };
 
 function loadEventListeners(): void {
   document.addEventListener('loadSection-settings', initHangoutSettings);
+
+  settingsSectionElement?.addEventListener('click', handleHangoutSettingsClicks);
+  updateHangoutPasswordForm?.addEventListener('submit', updateHangoutPassword);
 };
 
 function initHangoutSettings(): void {
@@ -93,7 +102,7 @@ function initHangoutSettings(): void {
     ),
 
     membersLimitSlider: new SliderInput(
-      'member-limit-input',
+      'members-limit-input',
       'member',
       MIN_HANGOUT_MEMBERS_LIMIT,
       MAX_HANGOUT_MEMBERS_LIMIT,
@@ -101,22 +110,104 @@ function initHangoutSettings(): void {
     ),
   };
 
+  document.addEventListener('availability-step-input_valueChange', updateSettingsButtons);
+  document.addEventListener('suggestions-step-input_valueChange', updateSettingsButtons);
+  document.addEventListener('voting-step-input_valueChange', updateSettingsButtons);
+  document.addEventListener('members-limit-input_valueChange', updateSettingsButtons);
+
   renderHangoutSettings();
   LoadingModal.remove();
 };
 
 function renderHangoutSettings(): void {
+  if (!hangoutSettingsState.settingsSectionMutationObserverActive) {
+    initSettingsSectionMutationObserver();
+  };
+};
+
+async function handleHangoutSettingsClicks(e: MouseEvent): Promise<void> {
+  if (!(e.target instanceof HTMLButtonElement)) {
+    return;
+  };
+
+  if (e.target.id === 'stages-settings-apply-btn') {
+    await updateHangoutStages();
+    return;
+  };
+
+  if (e.target.id === 'stages-settings-reset-btn') {
+    resetStageSliderValues();
+    return;
+  };
+
+  if (e.target.id === 'progress-hangout-btn') {
+    await progressHangout();
+    return;
+  };
+
+  if (e.target.id === 'members-limit-settings-apply-btn') {
+    await updateHangoutMembersLimit();
+    return;
+  };
+
+  if (e.target.id === 'members-limit-settings-reset-btn') {
+    resetMembersLimitSliderValues();
+    return;
+  };
+
+  if (e.target.getAttribute('data-goTo')) {
+    navigateHangoutSections(e);
+    return;
+  };
+
+  if (e.target.id === 'settings-password-copy-btn') {
+    // TODO: implement
+    return;
+  };
+
+  if (e.target.id === 'hangout-settings-password-input-reveal-btn') {
+    // TODO: implement
+    return;
+  };
+
+  if (e.target.id === 'delete-hangout-password-btn') {
+    // TODO: implement
+  };
+};
+
+async function updateHangoutStages(): Promise<void> {
   // TODO: implement
 };
 
-function calculateStepMinimumSliderValue(stage: number, currentStage: number, stageControlTimestamp: number): number {
-  if (stage < currentStage) {
-    return Math.round(stage / dayMilliseconds);
+async function progressHangout(): Promise<void> {
+  // TODO: implement
+};
+
+async function updateHangoutMembersLimit(): Promise<void> {
+  // TODO: implement
+};
+
+async function updateHangoutPassword(): Promise<void> {
+  // TODO: implement
+};
+
+function initSettingsSectionMutationObserver(): void {
+  if (!settingsSectionElement) {
+    return;
   };
 
-  if (currentStage === stage) {
-    return Math.ceil((Date.now() - stageControlTimestamp) / dayMilliseconds);
-  };
+  const observer: MutationObserver = new MutationObserver((mutations: MutationRecord[]) => {
+    for (const mutation of mutations) {
+      if (mutation.attributeName === 'class' && settingsSectionElement.classList.contains('hidden')) {
+        resetSliderValues();
+        hangoutSettingsState.settingsSectionMutationObserverActive = false;
 
-  return MIN_HANGOUT_PERIOD_DAYS;
+        observer.disconnect();
+        break;
+      };
+    };
+  });
+
+  observer.observe(settingsSectionElement, { attributes: true, attributeFilter: ['class'], subtree: false });
+  hangoutSettingsState.settingsSectionMutationObserverActive = true;
 };